// client/api/model/updateProjectRequest.ts
/**
 * Smart Power Electric API
 * Smart Power Electric API Documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class UpdateProjectRequest {
    /**
    * Client id of the project
    */
    'clientId'?: string | null;
    /**
    * Name of the project
    */
    'name'?: string | null;
    /**
    * Description of the project
    */
    'description'?: string | null;
    /**
    * Location of the project
    */
    'location'?: string | null;
    /**
    * Start date of the project
    */
    'startDate'?: string | null;
    /**
    * End date of the project
    */
    'endDate'?: string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "clientId",
            "baseName": "clientId",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "location",
            "baseName": "location",
            "type": "string"
        },
        {
            "name": "startDate",
            "baseName": "startDate",
            "type": "string"
        },
        {
            "name": "endDate",
            "baseName": "endDate",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return UpdateProjectRequest.attributeTypeMap;
    }
}